view autocomplete with suggestion
