+++ src/components/PresentationComponents/Tables/Tables.tsx
@@ -41,12 +41,10 @@ import ButtonDropdownColumnSelector from '@/components/SelectorComponents/Button
 import { useTableCellStructure } from '@/hooks/useTableCellStructure';
 import { getHeaderColomnColor } from '@/utils/functions/getColorStyle';
 import { filtersDataSend } from '@/utils/functions/filtersDataSend';
-import { generateRowsData } from '@/utils/functions/generateRowsData';
-import { generateColumnDef } from '@/utils/functions/generateColumnDef';
 import { fetchVoyageOptionsAPI } from '@/fetch/voyagesFetch/fetchVoyageOptionsAPI';
 import { fetchEnslavedOptionsList } from '@/fetch/pastEnslavedFetch/fetchPastEnslavedOptionsList';
 import { fetchEnslaversOptionsList } from '@/fetch/pastEnslaversFetch/fetchPastEnslaversOptionsList';
-
+import useDataTableProcessingEffect from '@/hooks/useDataTableProcessingEffect';
 
 const Tables: React.FC = () => {
     const dispatch: AppDispatch = useDispatch();
@@ -75,7 +73,6 @@ const Tables: React.FC = () => {
     const { inputSearchValue } = useSelector(
         (state: RootState) => state.getCommonGlobalSearch
     );
-
 
     const { isChangeGeoTree } = useSelector(
         (state: RootState) => state.getGeoTreeData
@@ -191,37 +188,15 @@ const Tables: React.FC = () => {
         currentBlockName, styleNameRoute,
     ]);
 
-    useEffect(() => {
-        const tableFileName = checkPagesRouteForVoyages(styleNameRoute!)
-            ? tableFlatfileVoyages
-            : checkPagesRouteForEnslaved(styleNameRoute!)
-                ? tableFlatfileEnslaved
-                : checkPagesRouteForEnslavers(styleNameRoute!)
-                    ? tableFlatfileEnslavers
-                    : null;
-        if (data.length > 0) {
-            const finalRowData = generateRowsData(data, tableFileName!);
-
-
-            const newColumnDefs: ColumnDef[] = tablesCell.map(
-                (value: TableCellStructure) =>
-                    generateColumnDef(value, visibleColumnCells)
-            );
-            dispatch(setColumnDefs(newColumnDefs));
-            dispatch(setRowData(finalRowData as Record<string, any>[]));
-        } else {
-            dispatch(setRowData([]));
-        }
-        // Ensure to return undefined if there's no cleanup needed
-        return undefined;
-    }, [
+    // Call the custom hook to Process Table Data
+    useDataTableProcessingEffect(
         data,
         visibleColumnCells,
-        dispatch,
         tableFlatfileVoyages,
         tableFlatfileEnslaved,
-        tableFlatfileEnslavers, tablesCell
-    ]);
+        tableFlatfileEnslavers,
+        tablesCell,
+    );
 
     const defaultColDef = useMemo(
         () => ({
